fixfixfix

+++ go/src/isubata/app.go
@@ -552,7 +552,7 @@ func fetchUnread(c echo.Context) error {
 				"SELECT COUNT(*) as cnt FROM message WHERE channel_id = ? AND ? < id",
 				chID, lastID)
 		} else {
-			cnt, err = redis.Int(redisConn.Do("GET", chID)) // keyがないときはcntは0になる
+			cnt, err = redis.Int64(redisConn.Do("GET", chID)) // keyがないときはcntは0になる
 			if err != redis.ErrNil {
 				return err
 			}
